[REFACTOR] Moves `FeedCachePolicy` to its own file.
